[apirnd] MOVE CURAND GENERATION FROM RAMBO TO RANDOMNUMBERKERNEL
This considerably cleans up rambo - and eventually allows runtime choices of RNDGEN
